combine testcases
