+++ e2e/testcases/team/scope-changes-after-office-change.spec.ts
@@ -5,6 +5,7 @@ import { faker } from '@faker-js/faker'
 import { CLIENT_URL, CREDENTIALS, GATEWAY_HOST } from '../../constants'
 import {
   createPIN,
+  drawSignature,
   getAuthTokens,
   getToken,
   login,
@@ -68,6 +69,170 @@ const expectVersionCard = async (
   await expect(page.getByText('Online', { exact: true })).toBeVisible()
   await expect(page.getByText('OpenCRVS v2.1.0', { exact: true })).toBeVisible()
 }
+
+test('Scope changes after office change - user loses access when the office changes', async ({
+  browser
+}) => {
+  test.setTimeout(300_000)
+
+  const page = await browser.newPage()
+  let username = ''
+  let fullName = ''
+  let childName = ''
+  let trackingId = ''
+  let eventId = ''
+  let draftCountBeforeChange = 0
+
+  await test.step('Create a new registrar user and set up initial record and drafts', async () => {
+    const adminToken = await getToken(CREDENTIALS.NATIONAL_SYSTEM_ADMIN)
+    const client = createClient(
+      GATEWAY_HOST + '/events',
+      `Bearer ${adminToken}`
+    )
+
+    const name = {
+      firstname: faker.person.firstName(),
+      surname: faker.person.lastName()
+    }
+
+    fullName = `${name.firstname} ${name.surname}`
+    username = `${name.firstname[0]}.${name.surname}`
+      .toLowerCase()
+      .replace(/[^a-z0-9.]/g, '')
+
+    const offices = await getLocations('CRVS_OFFICE', adminToken)
+    const ibomboDistrictOfficeId = getIdByName(
+      offices,
+      'Ibombo District Office'
+    )
+
+    await client.user.create.mutate({
+      name,
+      role: 'LOCAL_REGISTRAR',
+      primaryOfficeId: ibomboDistrictOfficeId,
+      mobile: `07${faker.string.numeric(8)}`,
+      email: faker.internet.email(),
+      fullHonorificName: fullName,
+      device: 'web',
+      data: {}
+    })
+
+    await loginWithNewUser(page, username)
+
+    const { token, refreshToken } = await getAuthTokens(
+      username,
+      NEW_USER_PASSWORD
+    )
+
+    await page.goto(`${CLIENT_URL}?refreshToken=${refreshToken}`)
+    await page.waitForSelector('#pin-input, #appSpinner', {
+      state: 'visible'
+    })
+    await createPIN(page)
+    await page.goto(CLIENT_URL)
+
+    const declaration = await createDeclaration(
+      token,
+      undefined,
+      ActionType.DECLARE
+    )
+
+    trackingId = declaration.trackingId!
+    eventId = declaration.eventId
+    childName = formatV2ChildName(declaration.declaration)
+
+    await searchFromSearchBar(page, childName, true)
+    await expect(page.getByTestId('tracking-id-value')).toContainText(
+      trackingId
+    )
+
+    const draftNames: string[] = []
+    for (let i = 0; i < 3; i++) {
+      draftNames.push(await createDraft(page))
+    }
+
+    draftCountBeforeChange = await countDraftRows(page, draftNames.length)
+    expect(draftCountBeforeChange).toBe(draftNames.length)
+
+    for (const draftName of draftNames) {
+      await expect(
+        page.getByRole('button', { name: draftName, exact: true })
+      ).toBeVisible()
+    }
+  })
+
+  await test.step('Local administrator moves the user to Isamba District Office', async () => {
+    await logout(page)
+    await login(page, CREDENTIALS.LOCAL_SYSTEM_ADMIN)
+
+    await page.getByRole('button', { name: 'Organisation' }).click()
+    await page.getByRole('button', { name: 'Central' }).click()
+    await page.getByRole('button', { name: 'Ibombo' }).click()
+    await page.getByRole('button', { name: 'Ibombo District Office' }).click()
+    await expect(page.locator('#content-name')).toHaveText(
+      'Ibombo District Office'
+    )
+
+    await page.getByRole('button', { name: fullName }).click()
+    await expect(page.locator('#content-name')).toHaveText(fullName)
+
+    await page.locator('#sub-page-header-munu-button-dropdownMenu').click()
+    await page.getByText('Edit details').click()
+    await expect(page.getByText('Confirm details')).toBeVisible()
+
+    await page.getByTestId('change-button-primaryOfficeId').click()
+    await page.locator('#searchable-select-primaryOfficeId').click()
+    await page.locator('#primaryOfficeId').fill('Isamba')
+    await page.getByText('Isamba District Office').click()
+    await page.getByRole('button', { name: 'Continue' }).click()
+
+    await page.locator('#role').click()
+    await page.locator('#react-select-2-option-1').click()
+    await page.getByRole('button', { name: 'Continue' }).click()
+
+    const signButton = page.getByRole('button', { name: 'Sign', exact: true })
+    if (await signButton.isVisible()) {
+      await signButton.click()
+      await drawSignature(page, 'signature_canvas_element', false)
+      await page.getByRole('button', { name: 'Apply' }).click()
+    }
+
+    await page.getByRole('button', { name: 'Continue' }).click()
+    await page.getByRole('button', { name: 'Confirm' }).click()
+    await page.getByTestId('confirm_office_change').click()
+
+    await expect(page.getByTestId('office-link-value')).toHaveText(
+      'Isamba District Office'
+    )
+  })
+
+  await test.step('User can no longer find the record or drafts after the office change', async () => {
+    await logout(page)
+
+    const { refreshToken } = await getAuthTokens(username, NEW_USER_PASSWORD)
+    expect(refreshToken).toBeDefined()
+
+    await page.goto(`${CLIENT_URL}?refreshToken=${refreshToken}`)
+    await page.waitForSelector('#pin-input, #appSpinner', {
+      state: 'visible'
+    })
+
+    await searchFromSearchBar(page, trackingId, false)
+    await expect(
+      page.getByRole('button', { name: trackingId, exact: true })
+    ).not.toBeVisible()
+
+    await page.goto(CLIENT_URL)
+    const draftCountAfterChange = await countDraftRows(page, 0)
+    expect(draftCountAfterChange).toBe(0)
+    expect(draftCountBeforeChange).toBeGreaterThan(0)
+
+    await page.goto(`${CLIENT_URL}/events/${eventId}`)
+    await expect(
+      page.getByText(`No event or draft found with id: ${eventId}`)
+    ).toBeVisible({ timeout: 30_000 })
+  })
+})
 
 test('Scope changes after office and role changes', async ({ browser }) => {
   test.setTimeout(300_000)
